Copyright/DMCA / Fraud / Trademark Infringement / Possible Phishing
From: "Allen, Lehanna" <Lehanna.Allen@fusionconnect.com>

A domain hosted with Namecheap has stolen our company's website content
and appears to be attempting to collect information for a phishing scam.
For the safety of our customers, we respectfully request the affected
pages or domain to be removed.

Content stolen:
http://84.32.100.30.sslip.io/
http://84.32.100.30.sslip.io/speedtest

Stolen from:
https://www.speakeasy.net/
https://www.speakeasy.net/speedtest